Move Priority to arc.util
