+++ arc-core/src/arc/util/Priority.java
@@ -1,4 +1,4 @@
-package arc;
+package arc.util;
 
 /** Utility class for ordering elements by priority. */
 public enum Priority {
